Added isAlive state for both players and as a requirement for the game to be running
I have modified the Player and Player2 classes with isAlive bool set to
true as default.
GameEngine is monitoring for conditions to set this variable to false
and if both players get false, the game freezes/stops.
We can add a state/screen  after the game freezes/stops with the score
and some other info if we want to.